requirements.txt
asgiref==3.5.2
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.4
Django==4.1.2
django-cleanup==6.0.0
django-modeltranslation==0.18.5
django_debug_toolbar==3.8.1
exceptiongroup==1.1.3
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.7
outcome==1.2.0
Pillow==9.2.0
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-decouple==3.6
python-gettext==4.1
pytz==2022.5
requests==2.28.1
rsa==4.9
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.2
trio-websocket==0.10.3
tzdata==2022.5
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.5.0
wsproto==1.2.0
